removed loading animation added loading animation as component
