+++ src/components/common/loadingAnimation/styles.module.scss
@@ -1,0 +1,5 @@
+.circle{
+  width: 70px;
+  height: 70px;
+  padding: 10px;
+}
