+++ src/containers/main/InvestContainer/styles.module.scss
@@ -27,12 +27,6 @@
     width: 100%;
     padding: 0 0 30px 0;
   }
-
-  .circle{
-    width: 70px;
-    height: 70px;
-    padding: 10px;
-  }
 }
 
 .container_input {
